feat #25: 예산 추천 API 구현.

+++ src/test/java/com/moneylog/api/category/service/CategoryBatchServiceTest.java
@@ -31,12 +31,12 @@ class CategoryBatchServiceTest {
         //then
         List<Category> categories = categoryRepository.findAll();
         assertThat(categories).allSatisfy(category -> assertThat(category.getAverageRatio()).isNotNull());
-        assertThat(getSumAverageRatio(categories)).isLessThan(100.0);
+        assertThat(getSumAverageRatio(categories)).isLessThan(100L);
     }
 
     private static double getSumAverageRatio(List<Category> categories) {
         return categories.stream()
-                .mapToDouble(Category::getAverageRatio)
+                .mapToLong(Category::getAverageRatio)
                 .sum();
     }
 }
